cherry pick compatible spring boot
